datastore changed :fire:

+++ pepper_hp/modules/python/DataStore.py
@@ -59,8 +59,8 @@ class DataStore(object):
             self.meta['summaries'].add(summary_name)
             self.file_handler['{}/{}/{}'.format(self._summary_path_, summary_name, 'image')] = np.array(image, dtype=np.uint8)
             self.file_handler['{}/{}/{}'.format(self._summary_path_, summary_name, 'label')] = np.array(label, dtype=np.uint8)
-            self.file_handler['{}/{}/{}'.format(self._summary_path_, summary_name, 'position')] = np.array(position, dtype=np.uint32)
-            self.file_handler['{}/{}/{}'.format(self._summary_path_, summary_name, 'index')] = np.array(index, dtype=np.uint16)
+            self.file_handler['{}/{}/{}'.format(self._summary_path_, summary_name, 'position')] = np.array(position, dtype=np.int32)
+            self.file_handler['{}/{}/{}'.format(self._summary_path_, summary_name, 'index')] = np.array(index, dtype=np.int32)
             self.file_handler['{}/{}/{}'.format(self._summary_path_, summary_name, 'contig')] = contig_name
             self.file_handler['{}/{}/{}'.format(self._summary_path_, summary_name, 'region_start')] = region_start
             self.file_handler['{}/{}/{}'.format(self._summary_path_, summary_name, 'region_end')] = region_end
